refactor: rename org.eclipse.kura.nm.common -> org.eclipse.kura.nm
