+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/configuration/NMSettingsConverter.java
@@ -23,9 +23,9 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.eclipse.kura.configuration.Password;
-import org.eclipse.kura.nm.common.KuraIpStatus;
-import org.eclipse.kura.nm.common.NMDeviceType;
-import org.eclipse.kura.nm.common.NetworkProperties;
+import org.eclipse.kura.nm.KuraIpStatus;
+import org.eclipse.kura.nm.NMDeviceType;
+import org.eclipse.kura.nm.NetworkProperties;
 import org.freedesktop.dbus.types.UInt32;
 import org.freedesktop.dbus.types.Variant;
 import org.freedesktop.networkmanager.settings.Connection;
