+++ kura/test/org.eclipse.kura.nm.test/src/test/java/org/eclipse/kura/nm/NetworkPropertiesTest.java
@@ -10,7 +10,7 @@
  * Contributors:
  *  Eurotech
  *******************************************************************************/
-package org.eclipse.kura.nm.common;
+package org.eclipse.kura.nm;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -24,7 +24,7 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
-import org.eclipse.kura.nm.common.NetworkProperties;
+import org.eclipse.kura.nm.NetworkProperties;
 import org.freedesktop.dbus.errors.NotSupported;
 import org.junit.Test;
